debugging

+++ maintest/libft.h
@@ -6,7 +6,7 @@
 /*   By: svan-hoo <svan-hoo@student.42.fr>          +#+  +:+       +#+        */
 /*                                                +#+#+#+#+#+   +#+           */
 /*   Created: 2023/10/03 16:53:50 by svan-hoo          #+#    #+#             */
-/*   Updated: 2023/10/12 20:49:34 by svan-hoo         ###   ########.fr       */
+/*   Updated: 2023/10/13 18:23:24 by svan-hoo         ###   ########.fr       */
 /*                                                                            */
 /* ************************************************************************** */
 
@@ -15,6 +15,7 @@
 # include <stddef.h>
 # include <stdio.h>
 # include <stdlib.h>
+# include <unistd.h>
 
 int			ft_atoi(const char *c);
 void		ft_bzero(void *s, size_t n);
@@ -29,8 +30,8 @@ void		*ft_memcpy(void *dest, const void *src, size_t n);
 void		*ft_memmove(void *dest, const void *src, size_t n);
 void		*ft_memset(void *s, int c, size_t n);
 char		*ft_strchr(char *str, int c);
-int			ft_strlcat(char *dst, const char *src, int l);
-int			ft_strlcpy(char *dst, const char *src, int l);
+size_t		ft_strlcat(char *dst, const char *src, size_t l);
+size_t		ft_strlcpy(char *dst, const char *src, size_t l);
 size_t		ft_strlen(const char *c);
 int			ft_strncmp(const char *a, const char *b, size_t n);
 const char	*ft_strnstr(const char *haystack, const char *needle, size_t n);
